+++ src/main/java/racingcar/domain/Generator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface Generator {
     int nextValue();
