+++ src/test/java/racingcar/domain/ManualValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Arrays;
 import java.util.LinkedList;
